# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GoMindscout API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /mentors:
    get:
      operationId: listMentors
      tags: [mentors]
      summary: Search and filter approved mentors
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
        - name: language
          in: query
          schema: { type: string }
        - name: minPrice
          in: query
          schema: { type: number }
        - name: maxPrice
          in: query
          schema: { type: number }
        - name: minRating
          in: query
          schema: { type: number }
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [featured, rating, price_asc, price_desc, experience, newest]
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Paginated mentor list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MentorListResponse"
    post:
      operationId: createMentorProfile
      tags: [mentors]
      summary: Submit a new mentor profile (pending approval)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MentorProfileInput"
      responses:
        "201":
          description: Created mentor profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MentorProfile"
...
  /mentors/me:
    get:
      operationId: getMyMentorProfile
      tags: [mentors]
      summary: Get the authenticated mentor's own profile
      responses:
        "200":
          description: Own mentor profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MentorProfile"
        "404":
          description: No mentor profile found
    patch:
      operationId: updateMyMentorProfile
      tags: [mentors]
      summary: Update own mentor profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MentorProfileUpdate"
      responses:
        "200":
          description: Updated mentor profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MentorProfile"
    delete:
      operationId: deleteMyMentorProfile
      tags: [mentors]
      summary: Delete the authenticated mentor's profile, packages, availability and all related data
      responses:
        "204":
          description: Mentor profile deleted
        "401":
          description: Unauthorized
        "404":
          description: No mentor profile found
...
  /mentors/{mentorId}:
    get:
      operationId: getMentor
      tags: [mentors]
      summary: Get a mentor profile by ID
      parameters:
        - name: mentorId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Mentor profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MentorProfile"
        "404":
          description: Not found

  # ── Packages ───────────────────────────────────────────────────────────────
...
  /admin/mentors:
    get:
      operationId: adminListMentors
      tags: [admin]
      summary: Admin list all mentors (any status)
      parameters:
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: All mentor profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MentorProfile"
...
  /admin/mentors/{mentorId}:
    delete:
      operationId: adminDeleteMentor
      tags: [admin]
      summary: Hard delete a mentor profile and all related data
      parameters:
        - name: mentorId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Mentor deleted
        "404":
          description: Mentor not found
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: Get current user's notifications
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
        "401":
          description: Unauthorized
  /notifications/read-all:
    post:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success: { type: boolean }
        "401":
          description: Unauthorized
  /notifications/{notificationId}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a single notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "401":
          description: Unauthorized
        "404":
          description: Not found
components:
  schemas:
...
    MentorProfile:
      type: object
      required: [id, userId, headline, status, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        fullName: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        headline: { type: string }
        bio: { type: ["string", "null"] }
        categoryId: { type: ["integer", "null"] }
        categoryName: { type: ["string", "null"] }
        industry: { type: ["string", "null"] }
        expertiseTags: { type: array, items: { type: string } }
        yearsExperience: { type: ["integer", "null"] }
        languages: { type: array, items: { type: string } }
        hourlyRate: { type: ["number", "null"] }
        currency: { type: ["string", "null"] }
        introVideoUrl: { type: ["string", "null"] }
        linkedinUrl: { type: ["string", "null"] }
        calendlyUrl: { type: ["string", "null"] }
        experiences: { type: ["array", "null"], items: { $ref: "#/components/schemas/MentorExperience" } }
        honorsAwards: { type: ["array", "null"], items: { $ref: "#/components/schemas/MentorHonorAward" } }
        publications: { type: ["array", "null"], items: { $ref: "#/components/schemas/MentorPublication" } }
        certifications: { type: ["array", "null"], items: { $ref: "#/components/schemas/MentorCertification" } }
        status: { type: string, enum: [pending, approved, rejected, suspended] }
        isFeatured: { type: boolean }
        averageRating: { type: ["number", "null"] }
        totalReviews: { type: integer }
        totalSessions: { type: integer }
        createdAt: { type: string }
    MentorExperience:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        company: { type: string }
        location: { type: ["string", "null"] }
        startDate: { type: ["string", "null"] }
        endDate: { type: ["string", "null"] }
        isCurrent: { type: boolean }
        description: { type: ["string", "null"] }
    MentorHonorAward:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        issuer: { type: ["string", "null"] }
        date: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
    MentorPublication:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        publisher: { type: ["string", "null"] }
        url: { type: ["string", "null"] }
        date: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
    MentorCertification:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        issuer: { type: ["string", "null"] }
        issueDate: { type: ["string", "null"] }
        expiryDate: { type: ["string", "null"] }
        credentialId: { type: ["string", "null"] }
    MentorProfileInput:
      type: object
      required: [headline]
      properties:
        headline: { type: string }
        bio: { type: string }
        categoryId: { type: integer }
        industry: { type: string }
        expertiseTags: { type: array, items: { type: string } }
        yearsExperience: { type: integer }
        languages: { type: array, items: { type: string } }
        hourlyRate: { type: number }
        currency: { type: string }
        introVideoUrl: { type: string }
        linkedinUrl: { type: string }
        calendlyUrl: { type: string }
        experiences: { type: array, items: { $ref: "#/components/schemas/MentorExperience" } }
        honorsAwards: { type: array, items: { $ref: "#/components/schemas/MentorHonorAward" } }
        publications: { type: array, items: { $ref: "#/components/schemas/MentorPublication" } }
        certifications: { type: array, items: { $ref: "#/components/schemas/MentorCertification" } }
    MentorProfileUpdate:
      type: object
      properties:
        headline: { type: string }
        bio: { type: string }
        categoryId: { type: integer }
        industry: { type: string }
        expertiseTags: { type: array, items: { type: string } }
        yearsExperience: { type: integer }
        languages: { type: array, items: { type: string } }
        hourlyRate: { type: number }
        introVideoUrl: { type: string }
        linkedinUrl: { type: string }
        calendlyUrl: { type: string }
        experiences: { type: array, items: { $ref: "#/components/schemas/MentorExperience" } }
        honorsAwards: { type: array, items: { $ref: "#/components/schemas/MentorHonorAward" } }
        publications: { type: array, items: { $ref: "#/components/schemas/MentorPublication" } }
        certifications: { type: array, items: { $ref: "#/components/schemas/MentorCertification" } }
    MentorListResponse:
      type: object
      required: [mentors, total, page, limit]
      properties:
        mentors:
          type: array
          items:
            $ref: "#/components/schemas/MentorProfile"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
...
    Notification:
      type: object
      required: [id, userId, type, title, message, isRead, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        type: { type: string }
        title: { type: string }
        message: { type: string }
        link: { type: ["string", "null"] }
        isRead: { type: boolean }
        createdAt: { type: string }
...